Converter: Wrapperコンポーネントの作成
