+++ apps/media-converter/src/app/Converter/index.tsx
@@ -1,1 +1,1 @@
-export * from "./Converter"
+export * from './ConverterWrapper';
