+++ apps/media-converter/src/index.tsx
@@ -1,1 +1,1 @@
-export * from "./app"
+export * from './app/Converter';
